Separate file for StitchAISpacing_V1
